% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{qsagesmetadata}
\alias{qsagesmetadata}
\title{Query SAGES codebook}
\usage{
qsagesmetadata(
  qterm,
  gt_table = FALSE,
  searchfields = c("variablefieldname", "fieldlabel",
    "choicescalculationsorsliderlabel", "fieldnote", "variablefieldname_3_99",
    "choicescalculationssliderlabels"),
  codebook_location = "https://quantsci.s3.amazonaws.com/Work/SAGES/SAGES_metadata.csv"
)
}
\arguments{
\item{qterm}{string The term to search for in the codebook.}

\item{gt_table}{logical Should the results be returned as a tibble or in a gt() table?}

\item{searchfields}{string The column names of the codebook to search.}

\item{codebook_location}{string The location of the codebook.}
}
\value{
A tibble containing the rows of the codebook that contain the search term.
}
\description{
This function searches the SAGES codebook for the specified query terms.  Its
  goal is to make it easier to find variables of interest, and to find which
  data files contain those variables.
}
\examples{
qsagesmetadata("ID")
}
